Code quality with "golangci-lint"
To ensure a good code quality the Go ecosystem has hundreds of linters,
each with a different purpose. Instead of installing and running
multiple linters separately "golangci-lint" [1] provides a uniform
interface to run most popular and useful linters in parallel and with
many additional configuration features.
The actual runner is open source [2] and can be used locally as well in
any private CI/CD pipelines. In order to use it for tmpl-go,
a `.golangci.yml` configuration file [3] has been added.

[1]: https://golangci-lint.run
[2]: https://github.com/golangci/golangci-lint
[3]: https://golangci-lint.run/usage/configuration

GH-4