Fix YAML multiline-string without "block chomping" for go-header linter
The configuration of the go-header [1] linter is defined in the
golangci-lint [2] YAML configuration, but the YAML multiline-string [3]
doesn't used "block chomping [4] which resulted in a final newline at
the end of the template.

```diff
-This source code is licensed under the MIT license found in the LICENSE file.\n
+This source code is licensed under the MIT license found in the LICENSE file.
```

This causes golangci-lint to fail because the configured template
content doesn't matched the parsed text.
To fix this problem the YAML "block chomping" syntax is now used for the
multiline-string so that the final newline at the end gets stripped.

[1]: https://github.com/denis-tingajkin/go-header
[2]: https://github.com/golangci/golangci-lint
[3]: https://yaml-multiline.info
[4]: https://yaml.org/spec/1.2/spec.html#id2794534

GH-6