Adapt CI workflow "on" run configurations from "tmpl" template repository
Before the CI workflow only used the `push` configuration for the `on`
field. To improve the performance a more fine grained configuration is
now used that has already been defined in the "tmpl" template
repository [1]:

- Only runs on pushes to the `main` branch.
- Only runs on pushes for `v*` tags.
- Always runs for pushes to PRs.

These changes help to keep the required GitHub Action run minutes for
the account of this repository as small as possible without wasting
resources for unnecessary runs.

[1]: https://github.com/svengreb/tmpl

GH-18