Rename "/app" to "/apps"
Using the majority `apps` as name for the `/app` directory [1] conveys a
better understanding of the actual use case and makes it more clear that
it can contain more than one application in the monorepo [2] layout.
This also aligns with the example in the `/web` directory
documentation [3] and the Yarn/NPM workspace configuration in the
`package.json` file [4] where both using the `apps` directory.

[1]: https://github.com/svengreb/tmpl-go/tree/main/app
[2]: https://trunkbaseddevelopment.com/monorepos
[3]: https://github.com/svengreb/tmpl-go/tree/main/web#example
[4]: https://github.com/svengreb/tmpl-go/blob/main/package.json#L24

GH-37