`golangci-lint` fails to run due to `revive`s unknown `time-equal` rule
The `time-equal` rule [1] was added in GH-64 [2], but the `revive`
linter [3] used by `golangci-lint` does not yet include the rule in the
current release version [4] but only merged it into the `main` branch.
Therefore linting failed because the rule is not known (yet) and
`golangci-lint` exits before running any linter.
To fix the problem the `time-equal` rule has been disabled again for now
and will be enabled again when it is available in a new `revive` version
that is used by `golangci-lint`.

[1]: https://github.com/mgechev/revive/blob/master/RULES_DESCRIPTIONS.md#time-equal
[2]: https://github.com/svengreb/tmpl-go/issues/64
[3]: https://github.com/mgechev/revive
[4]: https://github.com/mgechev/revive/releases/tag/v1.1.2

GH-76